[feat] Rename AccountStore to SettingsStore + add fetch method using url encoded data
