+++ src/plugins/preact-polyglot.js
@@ -22,7 +22,7 @@ const init = function (context, lang) {
       polyglot.extend(dict)
       polyglot.locale(lang)
     } catch (e) {
-      console.warning(`The dict phrases for "${lang}" can't be loaded`)
+      console.warn(`The dict phrases for "${lang}" can't be loaded`)
     }
   }
